# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Clean Architecture API with Better Auth integration",
        "title": "Next-Go-PG API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get statistics for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stats_interfaces_http.UserStatsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/stats_interfaces_http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Modify a stat value and broadcast the change via SSE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update user statistics",
                "parameters": [
                    {
                        "description": "Stat update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stats_interfaces_http.UpdateStatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stats_interfaces_http.UserStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/stats_interfaces_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/stats_interfaces_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "stats_interfaces_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
        "stats_interfaces_http.UpdateStatRequest": {
            "type": "object",
            "properties": {
                "delta": {
                    "type": "integer",
                    "example": 1
                },
                "field": {
                    "type": "string",
                    "example": "projects"
                }
            }
        },
        "stats_interfaces_http.UserStatsResponse": {
            "type": "object",
            "properties": {
                "activityToday": {
                    "type": "integer"
                },
                "lastLogin": {
                    "type": "string"
                },
                "memberSince": {
                    "type": "string"
                },
                "notifications": {
                    "type": "integer"
                },
                "projectCount": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}